Renamed dataclayobject module
