+++ src/dataclay/__init__.py
@@ -2,7 +2,7 @@
 Public dataclay functions exported to use (friendly) "from dataclay import ..."
 """
 from dataclay.runtime.Initializer import *
-from dataclay.DataClayObject import DataClayObject
+from dataclay.dataclay_object import DataClayObject
 from dataclay.DataClayObjMethods import dclayEmptyMethod, dclayMethod
 
 StorageObject = DataClayObject
